requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiohttp-retry==2.9.1
aiosignal==1.3.2
alembic==1.14.1
aniso8601==10.0.0
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
bcrypt==4.2.1
blinker==1.9.0
cachetools==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
Faker==15.3.2
Flask==3.0.2
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-Migrate==4.0.7
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
frozenlist==1.5.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.161.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.67.0
greenlet==3.1.1
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
Mako==1.3.9
MarkupSafe==3.0.2
multidict==6.1.0
nulltype==2.3.1
openai==1.13.3
plaid-python==19.0.0
propcache==0.2.1
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
requests==2.32.3
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
SQLAlchemy-serializer==1.4.1
tqdm==4.67.1
twilio==9.0.2
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
Werkzeug==3.1.3
yarl==1.18.3
